requirements.txt
# Core dependencies
click>=8.1.0
jinja2>=3.1.0
pyyaml>=6.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# CLI enhancements
rich>=13.0.0
questionary>=2.0.0

# Optional
gitpython>=3.1.0

# AI / LLM integration
openai>=1.12.0
tenacity>=8.2.0  # For retry logic

# Development tools
ruff>=0.1.0  # Fast Python linter
